[refactor] extract reified constructor cache
