+++ core/src/main/java/org/jruby/javasupport/Java.java
@@ -48,9 +48,7 @@ import java.lang.reflect.Modifier;
 import java.lang.reflect.Proxy;
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -623,33 +621,6 @@ public class Java implements Library {
         subclass.addMethod(context, "__jcreate!", new JCreateMethod(subclassSingleton));
     }
 
-    /**
-     * Used for concrete reified classes. Constructed in generated code (RubyClass)
-     */
-    public static class JCtorCache implements CallableSelector.CallableCache<ParameterTypes> {
-
-        private final NonBlockingHashMapLong<ParameterTypes> cache = new NonBlockingHashMapLong<>(8);
-        public final JavaConstructor[] constructors;
-        private final List<JavaConstructor> constructorList;
-
-        public JCtorCache(JavaConstructor[] constructors) {
-            this.constructors = constructors;
-            constructorList = Arrays.asList(constructors);
-        }
-
-        public int indexOf(JavaConstructor ctor) {
-            return constructorList.indexOf(ctor);
-        }
-
-        public final ParameterTypes getSignature(int signatureCode) {
-            return cache.get(signatureCode);
-        }
-
-        public final void putSignature(int signatureCode, ParameterTypes callable) {
-            cache.put(signatureCode, callable);
-        }
-    }
-
     public static class JCreateMethod extends JavaMethodN implements CallableSelector.CallableCache<JavaProxyConstructor> {
 
         private final NonBlockingHashMapLong<JavaProxyConstructor> cache = new NonBlockingHashMapLong<>(8);
@@ -665,7 +636,7 @@ public class Java implements Library {
          * @param runtime
          * @return Index of ctor in cache to call, or throws a new exception
          */
-        public static int forTypes(Ruby runtime, IRubyObject[] args, JCtorCache cache) {
+        public static int forTypes(Ruby runtime, IRubyObject[] args, ConstructorCache cache) {
             JavaConstructor ctor = matchConstructorIndex(runtime.getCurrentContext(), cache.constructors, cache,
                     args.length, args);
             int index = cache.indexOf(ctor);
